refactor(cli): extract info batch into commands/info.ts (5.0-alpha)
Thirteenth cut of the cli.ts god-module split — first of the independent leaf
batches now that the entangled clusters are all out.

- New src/commands/info.ts — the read-only status/inspection commands, all
  exported: cmdStatus, cmdHealth, cmdIngest, cmdSupplyChain, cmdAgentAudit,
  cmdWhoami, cmdVersion. No cross-cluster cmd* calls; cmdVersion sources
  KIT_VERSION from the cli-checks-shared seam, cmdHealth reuses buildHealthCtx
  from cli-shared.
- cli.ts imports them back; COMMANDS table unchanged. Pruned now-unused imports
  (gatherStatus, quickSubsystems, resolveMode/modeScore, buildHealthCtx,
  existsSync, kitConfig type).

Verification: tsc clean, eslint 0 errors, full suite 2580 passing, prettier
clean, runtime dispatch smoke-tested (version/status/supply-chain).

cli.ts: 2119 -> 1853 lines (-266; ~71% smaller than the 6330-line original).
